requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
Pillow==8.3.0
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
twilio==6.61.0
urllib3==1.26.6
whitenoise==5.2.0
